sisserad/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-jazzmin==3.0.0
djangorestframework==3.15.1
sqlparse==0.5.0
typing_extensions==4.12.0
